@@ -4,17 +4,19 @@ import { createScheduler } from "../../testing/rxjs";
 describe("fallback operator", () => {
 
     it.each([{
-        in$: " ----------|",
-        out$: "---f--f--f|"
+        in$:     "----------|",
+        timeout: "---|",
+        out$:    "---f--f--f|"
     }, {
-        in$: " a----b----------c-----|",
-        out$: "a--f-b--f--f--f-c--f--|"
-    }])("works for case %#", ({ in$, out$ }) => {
+        in$:     "a----b----------c-----|",
+        timeout: "---|",
+        out$:    "a--f-b--f--f--f-c--f--|"
+    }])("works for case %#", ({ in$, out$, timeout }) => {
         const run = createScheduler();
         run(({ hot, time, expectObservable }) => {
             let source$ = hot(in$);
             let expected$ = out$;
-            let actual$ = source$.pipe(fallback("f", time("---|")))
+            let actual$ = source$.pipe(fallback("f", time(timeout)))
     
             expectObservable(actual$).toBe(expected$);
         })
